+++ core/org.osate.core.tests/src/org/osate/core/tests/issues/Issue3109Test.java
@@ -98,6 +98,6 @@ public class Issue3109Test extends XtextTest {
 				.toList();
 		assertEquals(connection.getInstanceObjectPath(), 1, associations.size());
 		var expression = associations.getFirst().getOwnedValues().getFirst().getOwnedValue();
-		return (long) ((IntegerLiteral) expression).getValue();
+		return ((IntegerLiteral) expression).getValue();
 	}
 }
